requirements.txt
asgiref==3.8.1
Cerberus==1.3.5
Django==5.0.4
djangorestframework==3.15.1
greenlet==3.0.3
inflect==7.3.1
isort==5.13.2
more-itertools==10.3.0
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
sqlparse==0.5.0
typeguard==4.3.0
typing_extensions==4.11.0
